pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "figment-agent"
version = "0.1.0"
description = "E100 expansion list pipeline: Looker, Enterpret Wisdom (MCP), deterministic merge/score"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27",
    "requests>=2.31",
    "gspread>=6.0",
    "google-auth>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
include = ["agents*", "core*", "outputs*", "bootstrap*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
